fix: Delete action dir
