+++ frontend/src/reducer/index.js
@@ -1,8 +1,8 @@
-import signup from './signup'
+import { inputField } from './signup'
 import { combineReducers } from 'redux'
 
 const rootReducers = combineReducers({
-    signup: signup
+    signup: inputField
 })
 
 export default rootReducers
